+++ js/effects.js
@@ -1,56 +1,9 @@
+import { EFFECTS } from './constants.js';
+
 const imgPreview = document.querySelector('.img-upload__preview img');
 const uploadForm = document.querySelector('.img-upload__form');
 const effectLevel = document.querySelector('.effect-level__value');
 const sliderElement = document.querySelector('.effect-level__slider');
-
-const EFFECTS = [
-  {
-    name: 'none',
-    min: 0,
-    max: 100,
-    step: 1,
-  },
-  {
-    name: 'chrome',
-    style: 'grayscale',
-    min: 0,
-    max: 1,
-    step: 0.1,
-    unit: '',
-  },
-  {
-    name: 'sepia',
-    style: 'sepia',
-    min: 0,
-    max: 1,
-    step: 0.1,
-    unit: '',
-  },
-  {
-    name: 'marvin',
-    style: 'invert',
-    min: 0,
-    max: 100,
-    step: 1,
-    unit: '%',
-  },
-  {
-    name: 'phobos',
-    style: 'blur',
-    min: 0,
-    max: 3,
-    step: 0.1,
-    unit: 'px',
-  },
-  {
-    name: 'heat',
-    style: 'brightness',
-    min: 1,
-    max: 3,
-    step: 0.1,
-    unit: '',
-  },
-];
 
 const DEFAULT_EFFECT = EFFECTS[0];
 let currentEffect = DEFAULT_EFFECT;
@@ -109,7 +62,14 @@ noUiSlider.create(sliderElement, {
 });
 updateSlider();
 
-uploadForm.addEventListener('change', onFormChange);
+const addChangeEvent = () => {
+  uploadForm.addEventListener('change', onFormChange);
+};
+
+const removeChangeEvent = () => {
+  uploadForm.removeEventListener('change', onFormChange);
+};
+
 sliderElement.noUiSlider.on('update', onSliderUpdate);
 
-export {resetEffect};
+export { resetEffect, addChangeEvent, removeChangeEvent };
